feat: Pass context to Serve
If things are added to the context (e.g. a logger) and we want to
access it in a Receiver we have to wrap the receiver in something that
passes along the original context which is cumbersome and easy to
forget.

By passing context into Serve, we can maintain a much more natural flow
of contexts from a listener, to the Server, and then to the receiver,
which will preserve any information that was added to the context at
the beginning of the flow.